# docs/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Sentinel AML API",
        "description": "Transaction monitoring: ingestion, rule-based detection, alerts and case management. Roles: ANALYST (queue & disposition), SUPERVISOR (+ full PII, case closure, reopen), ADMIN (+ ingestion, rule/watchlist/FX configuration).",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "basic": []
        }
    ],
...
    "paths": {
        "/api/v1/fx-rates/{currency}": {
            "put": {
                "tags": [
                    "Reference data"
                ],
                "summary": "Create or update the rate to the base currency (applies to transactions ingested afterwards)",
                "description": "Guardrail: a move of more than 10% is rejected with 422 FX_RATE_DEVIATION unless confirmLargeChange=true",
                "operationId": "setRate",
                "parameters": [
                    {
                        "name": "currency",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "pattern": "[A-Z]{3}",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RateUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExchangeRate"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/watchlist": {
            "get": {
                "tags": [
                    "Reference data"
                ],
                "operationId": "watchlist",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WatchlistEntry"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reference data"
                ],
                "operationId": "add",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewWatchlistEntry"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WatchlistEntry"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/ingest/transactions": {
            "post": {
                "tags": [
                    "Ingestion"
                ],
                "summary": "Bulk load transactions from a JSON array, then run detection in parallel (detect=false to skip)",
                "operationId": "transactions",
                "parameters": [
                    {
                        "name": "detect",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": true
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/TransactionRecord"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/IngestionResult"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/watchlist/{id}": {
            "patch": {
                "tags": [
                    "Reference data"
                ],
                "summary": "Activate / deactivate an entry (entries are never deleted, for auditability)",
                "operationId": "toggle",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WatchlistToggle"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WatchlistEntry"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "RateUpdate": {
                "required": [
                    "rateToBase"
                ],
                "type": "object",
                "properties": {
                    "rateToBase": {
                        "type": "number"
                    },
                    "confirmLargeChange": {
                        "type": "boolean"
                    }
                }
            },
            "ExchangeRate": {
                "type": "object",
                "properties": {
                    "currency": {
                        "type": "string"
                    },
                    "rateToBase": {
                        "type": "number"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedBy": {
                        "type": "string"
                    }
                }
            },
            "NewWatchlistEntry": {
                "required": [
                    "entryType",
                    "listName",
                    "value"
                ],
                "type": "object",
                "properties": {
                    "entryType": {
                        "pattern": "COUNTRY|COUNTERPARTY",
                        "type": "string"
                    },
                    "value": {
                        "maxLength": 200,
                        "minLength": 0,
                        "type": "string"
                    },
                    "listName": {
                        "maxLength": 60,
                        "minLength": 0,
                        "type": "string"
                    },
                    "reason": {
                        "maxLength": 300,
                        "minLength": 0,
                        "type": "string"
                    }
                }
            },
            "WatchlistEntry": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "entryType": {
                        "type": "string"
                    },
                    "value": {
                        "type": "string"
                    },
                    "listName": {
                        "type": "string"
                    },
                    "reason": {
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "TransactionRecord": {
                "required": [
                    "accountNumber",
                    "amount",
                    "channel",
                    "currency",
                    "direction",
                    "occurredAt",
                    "txnRef"
                ],
                "type": "object",
                "properties": {
                    "txnRef": {
                        "maxLength": 60,
                        "minLength": 0,
                        "type": "string"
                    },
                    "accountNumber": {
                        "maxLength": 40,
                        "minLength": 0,
                        "type": "string"
                    },
                    "direction": {
                        "pattern": "CREDIT|DEBIT",
                        "type": "string"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "currency": {
                        "pattern": "[A-Z]{3}",
                        "type": "string"
                    },
                    "counterpartyName": {
                        "maxLength": 200,
                        "minLength": 0,
                        "type": "string"
                    },
                    "counterpartyAccount": {
                        "maxLength": 60,
                        "minLength": 0,
                        "type": "string"
                    },
                    "counterpartyCountry": {
                        "pattern": "[A-Z]{2}",
                        "type": "string"
                    },
                    "channel": {
                        "pattern": "CASH|WIRE|NEFT|RTGS|IMPS|UPI|CARD|ATM|CHEQUE|SWIFT",
                        "type": "string"
                    },
                    "jurisdiction": {
                        "pattern": "[A-Z]{2}",
                        "type": "string"
                    },
                    "occurredAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "IngestionResult": {
                "type": "object",
                "properties": {
                    "batchId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "recordType": {
                        "type": "string"
                    },
                    "received": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "accepted": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "duplicates": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "rejected": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "errors": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RowError"
                        }
                    },
                    "errorsTruncated": {
                        "type": "boolean"
                    },
                    "ingestMs": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "detection": {
                        "$ref": "#/components/schemas/Summary"
                    }
                }
            },
            "RowError": {
                "type": "object",
                "properties": {
                    "line": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "key": {
                        "type": "string"
                    },
                    "error": {
                        "type": "string"
                    }
                }
            },
            "Summary": {
                "type": "object",
                "properties": {
                    "evaluated": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "skipped": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "transactionsWithHits": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "alertsCreated": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "alertsUpdated": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "failures": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "elapsedMs": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "txnPerSecond": {
                        "type": "number",
                        "format": "double"
                    }
                }
            },
...
            "WatchlistToggle": {
                "required": [
                    "active"
                ],
                "type": "object",
                "properties": {
                    "active": {
                        "type": "boolean"
                    },
                    "reason": {
                        "maxLength": 300,
                        "minLength": 0,
                        "type": "string"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "basic": {
                "type": "http",
                "scheme": "basic"
            }
        }
    }
}